+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 * New `IsLogMasked` option for masking entity properties in logs
+* Dependabot scaffolding. Can be excluded with `IncludeDependabot = false` at the domain template level
+* Github test action scaffolding. Can be excluded with `IncludeGithubTestActions = false` at the api template level
 
 ### Updated